configs: set SPL_TEXT_BASE by default for k3 platforms
SPL_TEXT_BASE is used as the load address for the main domain SPL on k3
platforms.

Since the config value is the same for every board, this patch sets the
value 0x80080000 as default for all 64-bit ARCH_K3, 0x43c00000 as
default for the R5 cores and deletes the instances of SPL_TEXT_BASE in
individual defconfigs.

Signed-off-by: Anshul Dalal <anshuld@ti.com>
Reviewed-by: Tom Rini <trini@konsulko.com>
---
Changes in v3:
 * Added Reviewed-by
 * Moved SPL_TEXT_BASE for phycore_am62x and j784s4
v2:
https://lore.kernel.org/u-boot/20250410135401.946596-1-anshuld@ti.com/
Changes in v2:
 * Moved SPL_TEXT_BASE for r5 cores
v1:
https://lore.kernel.org/u-boot/20250409124739.3843126-1-anshuld@ti.com/
Signed-off-by: Anshul Dalal <anshuld@ti.com>