arch: arm: k3-binman: add fit for falcon boot
This adds creation of tifalcon.bin for the am62a, 62p and 62x.

The contents are the same as the existing tispl.bin but A53's spl and
the fdt have been removed as they are not needed in falcon boot.

This reduces boot time since the payload size is smaller and we don't
need to authenticate the spl and fdt in secure boot.

Signed-off-by: Anshul Dalal <anshuld@ti.com>