mach-k3: mmu: add dynamic carveouts for MMU table
Currently in upstream u-boot we only provide a single MMU table for
all k3 platforms, this does not scale for devices with reserved memory
outside the range 0x9e780000 - 0xa0000000 (eg j722s[1]) or for devices
with < 2GiB of memory (eg am62sip).

To properly configure the MMU on various k3 platforms, the
reserved-memory regions need to be queried at runtime from the
device-tree and the MMU table should be updated accordingly.

This patch adds the required fixups to the MMU table (during proper
U-boot stage) by marking the reserved regions as non cacheable and
the remaining area as cacheable.

For the A-core SPL, the DDR is divided into three areas as follows:

  0x80000000  +------------------+ CFG_SYS_SDRAM_BASE
              |                  |
    512KiB    | #1 Non cacheable | reserved for ATF/OP-TEE
              |                  |
  0x80080000  |------------------| CONFIG_SPL_TEXT_BASE
              |                  |
    128MiB    | #2 Cacheable     | for runtime usage by the A-core SPL
              |                  |
  0x88080000  |------------------|
              |                  |
    1.9GiB    | #3 Non cacheable | reserved for DM/tifs-stub etc.
              |                  |
  0x100000000 +------------------+ End of DDR

The 128MiB size is chosen to allow for future use cases such as A53 SPL
falcon boot which would require loading kernel image from the SPL stage
while also ensuring the reserved memory regions that all exist past
0x88080000 are non cacheable.

[1]:
https://git.ti.com/cgit/ti-u-boot/ti-u-boot/tree/arch/arm/mach-k3/arm64/arm64-mmu.c?h=ti-u-boot-2025.01-next#n54

Depends on:
  [PATCH v1] ti: k3: abstract common fdt api for reserved mem fixups
  https://lore.kernel.org/u-boot/20250522150941.563959-1-anshuld@ti.com/

Boot test:
  http://lcpdresults.itg.ti.com/launcher/results/28618
  (*internal link*: will be removed before posting upstream)

Signed-off-by: Anshul Dalal <anshuld@ti.com>