configs: add falcon defconfigs for am62,62a,62p
The SPL_FS_LOAD_ARGS_NAME config is used for the arguments to the kernel
(dtb in our case) in falcon boot.

So, it can't be set in a generic config fragment. Therefore we add new
falcon defconfigs for all supported platforms that include the base
defconfigs and the falcon fragment while setting their corresponding
dtbs.

Signed-off-by: Anshul Dalal <anshuld@ti.com>