spl: falcon mode: load kernel before args file
Currently in falcon mode, certain boot media load the kernel image first
followed by the args whereas others load in the opposite order. This
inconsistency means falcon boot doesn't behave consistently across
various boot media.

For example, in the case where the kernel file is a FIT with the kernel
image present alongside the dtb and the args file is another DT, which
DT should be picked? The one form the FIT or the one set by the args
file? Currently this depends entirely on how the boot media handles
falcon mode.

Therefore, this patch enforces the load order of the kernel image first
followed by the args file. So in the above example, the dt from FIT
would take precedence.

Signed-off-by: Anshul Dalal <anshuld@ti.com>