ti: add speed grades support for AM62a
Hi all,

TI offers SoCs in various speed grades, each speed grade specifies a
certain maximum operating frequency of the clocks for each core.

In K3 boot flow, the R5 SPL starts the A53 or A72 core and configures
the correct clocks and power using the K3 ARM64 rproc driver
(compatible: ti,am654-rproc). However, the driver expects the dt node
for the ARM64 core to be set with a correct "assigned-clock-rates"
value.

Currently the dt has a value of 1.2GHz for the A53 core on AM62a, this
is incorrect for lower speed grades. Therefore this patch set adds
support for fixing this value at runtime based on the detected speed
grade from the efuse MMR.

For the speed grade table, refer to Table 6-1 of the AM62a datasheet[1].

[1]: https://www.ti.com/lit/ds/symlink/am62a7.pdf

To: u-boot@lists.denx.de
Cc: Tom Rini <trini@konsulko.com>
Cc: Aparna Patra <a-patra@ti.com>
Cc: Bryan Brattlof <bb@ti.com>
Cc: Udit Kumar <u-kumar1@ti.com>
Cc: Dhruva Gole <d-gole@ti.com>
Cc: Judith Mendez <jm@ti.com>
Cc: Beleswar Padhi <b-padhi@ti.com>
Cc: Andrew Davis <afd@ti.com>
Cc: Aniket Limaye <a-limaye@ti.com>
Cc: Vignesh Raghavendra <vigneshr@ti.com>
Cc: Nishanth Menon <nm@ti.com>
Signed-off-by: Anshul Dalal <anshuld@ti.com>

--- b4-submit-tracking ---
{
  "series": {
    "revision": 1,
    "change-id": "20251029-ti_speed_grade_fix-e7f3f6806878",
    "prefixes": [
      "uU"
    ]
  }
}