mach-k3: refactor A-Core frequency fixup
The K3 ARM64 rproc driver uses the "assigned-clock-rates" value in the
respective "/a72@0" or "/a53@0" node to properly configure the clocks
for the ARM64 core.

Although the clock value in the DT node might need to be fixed based on
SoC's speed grade etc. Certain SoCs such as AM62p and AM62x already had
this implemented, this patch moves the common code to common.c to avoid
reuse and simplify speed grade handling.

Signed-off-by: Anshul Dalal <anshuld@ti.com>